+++ lib/hop_tasks.dart
@@ -1,13 +1,12 @@
 library hop_tasks;
 
 import 'dart:async';
-import 'dart:collection';
 import 'dart:io';
 import 'dart:math' as math;
 
 import 'package:args/args.dart';
 import 'package:meta/meta.dart';
-import 'package:pathos/path.dart' as path;
+import 'package:path/path.dart' as pathos;
 import 'package:unittest/unittest.dart' as unittest;
 
 import 'package:bot/bot.dart';
@@ -19,7 +18,6 @@ import 'package:hop/hop.dart';
 part 'src/hop_tasks/bench_task.dart';
 part 'src/hop_tasks/copy_directory.dart';
 part 'src/hop_tasks/dartanalyzer.dart';
-part 'src/hop_tasks/dart_analyzer.dart';
 part 'src/hop_tasks/dart2js.dart';
 part 'src/hop_tasks/dartdoc.dart';
 part 'src/hop_tasks/git_tasks.dart';