+++ lib/src/hop/task_fail_error.dart
@@ -3,7 +3,7 @@ part of hop;
 class _TaskFailError extends Error {
   final String message;
 
-  const _TaskFailError(this.message);
+  _TaskFailError(this.message);
 
   String toString() => "TaskFailError: $message";
 }