+++ lib/src/hop_experimental.dart
@@ -1,4 +1,4 @@
-library tool.tasks.shared;
+library hop_experimental;
 
 import 'dart:async';
 import 'dart:io';
@@ -6,10 +6,8 @@ import 'dart:io';
 import 'package:html5lib/dom.dart' as dom;
 import 'package:html5lib/parser.dart';
 
-import 'package:bot/bot.dart';
-
 Future<bool> transformHtml(String filePath,
-    Func1<dom.Document, Future<dom.Document>> transformer) {
+    Future<dom.Document> transformer(dom.Document doc)) {
 
   return transformFile(filePath, (String content) {
     var parser = new HtmlParser(content, generateSpans: true);
@@ -23,7 +21,7 @@ Future<bool> transformHtml(String filePath,
 }
 
 Future<bool> transformFile(String filePath,
-    Func1<String, Future<String>> transformer) {
+    Future<String> transformer(String input)) {
 
   final file = new File(filePath);
   assert(file.existsSync());