+++ lib/src/hop_tasks/unit_test.dart
@@ -5,6 +5,7 @@ const _summaryFlag = 'summary';
 const _summaryAll = 'all';
 const _summaryFail = 'fail';
 const _summaryPass = 'pass';
+const _summaryError = 'error';
 
 Task createUnitTestTask(Action1<unittest.Configuration> unitTestAction) {
   return new Task.async((TaskContext ctx) {
@@ -17,7 +18,10 @@ Task createUnitTestTask(Action1<unittest.Configuration> unitTestAction) {
     final failSummary =
         (summaryFlag == _summaryAll || summaryFlag == _summaryFail);
 
-    final config = new _HopTestConfiguration(ctx, failSummary, passSummary);
+    final errorSummary =
+        (summaryFlag == _summaryAll || summaryFlag == _summaryError);
+
+    final config = new _HopTestConfiguration(ctx, failSummary, passSummary, errorSummary);
 
     // TODO: wrap this in a try/catch
     unitTestAction(config);
@@ -57,24 +61,22 @@ void _unittestParserConfig(ArgParser parser) {
       help: "Just list the test case names. Don't run them. Any filter is still applied.");
   parser.addOption(_summaryFlag, abbr: 's',
       help: 'Summarize the results of individual tests.',
-      allowed: [_summaryAll, _summaryFail, _summaryPass],
+      allowed: [_summaryAll, _summaryFail, _summaryPass, _summaryError],
       allowMultiple: false);
 }
 
-class _HopTestConfiguration implements unittest.Configuration {
+class _HopTestConfiguration extends unittest.Configuration {
   final Completer<bool> _completer = new Completer<bool>();
   final TaskContext _context;
   final bool failSummary;
   final bool passSummary;
+  final bool errorSummary;
 
-  _HopTestConfiguration(this._context, this.failSummary, this.passSummary);
+  _HopTestConfiguration(this._context, this.failSummary, this.passSummary,
+      this.errorSummary) : super.blank();
 
   Future<bool> get future => _completer.future;
 
-  @override
-  String get name => 'hop_tasks.unittest';
-
-  @override
   bool get autoStart => false;
 
   @override
@@ -146,6 +148,13 @@ ${testCase.stackTrace}''');
       });
     }
 
+    if(errorSummary) {
+      final summaryCtx = _context.getSubLogger('ERROR');
+      results.where((tc) => tc.result == unittest.ERROR).forEach((tc) {
+        summaryCtx.severe(tc.description);
+      });
+    }
+
     if(success) {
       _context.info(message);
     } else {
@@ -157,17 +166,8 @@ ${testCase.stackTrace}''');
   void onDone(bool success) {
     _completer.complete(success);
   }
-
-  @override
-  @deprecated
-  void handleExternalError(error, String message, [String stack = '']) {
-    // should never occur in the context of hop runner
-    if(unittest.currentTestCase == null) {
-      _context.severe('$message\nCaught $error');
-      _context.fail('An external error occured in the test suite outside of a Unit Test');
-    } else {
-      unittest.currentTestCase.error('An external error occured in the test suite during ${unittest.currentTestCase.description}\n$message\nCaught $error');
-    }
-  }
 }
 
+/** Indent each line in [str] by two spaces. */
+String _indent(String str) =>
+  str.replaceAll(new RegExp("^", multiLine: true), "  ");