+++ test/hop/_hop.dart
@@ -3,11 +3,8 @@ library test_hop;
 import 'dart:async';
 import 'dart:io';
 import 'package:bot/bot.dart';
-import 'package:bot_io/bot_io.dart';
-import 'package:bot/bot_test.dart';
 import 'package:hop/hop.dart';
 import 'package:logging/logging.dart';
-import 'package:meta/meta.dart';
 import 'package:unittest/unittest.dart';
 
 part 'arg_tests.dart';