+++ test/hop_tasks/_hop_tasks.dart
@@ -2,13 +2,10 @@ library test_hop_tasks;
 
 import 'dart:async';
 import 'dart:io';
-import 'dart:crypto';
-import 'package:pathos/path.dart' as pathos;
+import 'package:path/path.dart' as pathos;
 import 'package:unittest/unittest.dart';
-import 'package:bot/bot.dart';
 import 'package:bot_io/bot_git.dart';
 import 'package:bot_io/bot_io.dart';
-import 'package:bot/bot_test.dart';
 import 'package:hop/hop.dart';
 import 'package:hop/hop_tasks.dart';
 import '../hop/_hop.dart';
@@ -16,7 +13,6 @@ import '../hop/_hop.dart';
 part 'process_tests.dart';
 part 'git_tests.dart';
 part 'dartanalyzer_tests.dart';
-part 'dart_analyzer_tests.dart';
 part 'copy_directory_tests.dart';
 part 'compiler_tests.dart';
 
@@ -26,7 +22,6 @@ void main() {
     GitTests.register();
     DartAnalyzerTests.register();
     CopyDirectoryTests.register();
-    Dart_AnalyzerTests.register();
     CompilerTests.register();
   });
 }